+++ source/DSCResources/DSC_SqlRS/DSC_SqlRS.psm1
@@ -18,6 +18,15 @@ $script:localizedData = Get-LocalizedData -DefaultUICulture 'en-US'
 
     .PARAMETER DatabaseInstanceName
         Name of the SQL Server instance to host the Reporting Service database.
+
+    .PARAMETER Encrypt
+        Specifies how encryption should be enforced when using command `Invoke-SqlCmd`.
+        When not specified, the default value is `Mandatory`.
+
+        This value maps to the Encrypt property SqlConnectionEncryptOption
+        on the SqlConnection object of the Microsoft.Data.SqlClient driver.
+
+        This parameter can only be used when the module SqlServer v22.x.x is installed.
 #>
 function Get-TargetResource
 {
@@ -36,7 +45,12 @@ function Get-TargetResource
 
         [Parameter(Mandatory = $true)]
         [System.String]
-        $DatabaseInstanceName
+        $DatabaseInstanceName,
+
+        [Parameter()]
+        [ValidateSet('Mandatory', 'Optional', 'Strict')]
+        [System.String]
+        $Encrypt
     )
 
     Write-Verbose -Message (
@@ -53,6 +67,7 @@ function Get-TargetResource
         ReportsReservedUrl           = $null
         UseSsl                       = $false
         IsInitialized                = $false
+        Encrypt                      = $Encrypt
     }
 
     $reportingServicesData = Get-ReportingServicesData -InstanceName $InstanceName
@@ -170,6 +185,15 @@ function Get-TargetResource
         settings change. If this parameter is set to $true, Reporting Services
         will not be restarted, even after initialization.
 
+    .PARAMETER Encrypt
+        Specifies how encryption should be enforced when using command `Invoke-SqlCmd`.
+        When not specified, the default value is `Mandatory`.
+
+        This value maps to the Encrypt property SqlConnectionEncryptOption
+        on the SqlConnection object of the Microsoft.Data.SqlClient driver.
+
+        This parameter can only be used when the module SqlServer v22.x.x is installed.
+
     .NOTES
         To find out the parameter names for the methods in the class
         MSReportServer_ConfigurationSetting it's easy to list them using the
@@ -247,7 +271,12 @@ function Set-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $SuppressRestart
+        $SuppressRestart,
+
+        [Parameter()]
+        [ValidateSet('Mandatory', 'Optional', 'Strict')]
+        [System.String]
+        $Encrypt
     )
 
     $reportingServicesData = Get-ReportingServicesData -InstanceName $InstanceName
@@ -470,9 +499,14 @@ function Set-TargetResource
                 For more details, see the article [Connect to SQL Server with strict encryption](https://learn.microsoft.com/en-us/sql/relational-databases/security/networking/connect-with-strict-encryption?view=sql-server-ver16)
                 and [Configure SQL Server Database Engine for encrypting connections](https://learn.microsoft.com/en-us/sql/database-engine/configure-windows/configure-sql-server-encryption?view=sql-server-ver16).
             #>
-            if ((Get-Command -Name Invoke-SqlCmd).Parameters.Keys -contains 'Encrypt')
+            if ($PSBoundParameters.ContainsKey('Encrypt'))
             {
-                $invokeSqlCmdParameters.Encrypt = 'Optional'
+                $commandInvokeSqlCmd = Get-Command -Name 'Invoke-SqlCmd'
+
+                if ($null -ne $commandInvokeSqlCmd -and $commandInvokeSqlCmd.Parameters.Keys -contains 'Encrypt')
+                {
+                    $invokeSqlCmdParameters.Encrypt = $Encrypt
+                }
             }
 
             Invoke-SqlCmd @invokeSqlCmdParameters -Query $reportingServicesDatabaseScript.Script
@@ -848,7 +882,16 @@ function Set-TargetResource
     .PARAMETER SuppressRestart
         Reporting Services need to be restarted after initialization or
         settings change. If this parameter is set to $true, Reporting Services
-        will not be restarted, even after initialisation.
+        will not be restarted, even after initialization.
+
+    .PARAMETER Encrypt
+        Specifies how encryption should be enforced when using command `Invoke-SqlCmd`.
+        When not specified, the default value is `Mandatory`.
+
+        This value maps to the Encrypt property SqlConnectionEncryptOption
+        on the SqlConnection object of the Microsoft.Data.SqlClient driver.
+
+        This parameter can only be used when the module SqlServer v22.x.x is installed.
 #>
 function Test-TargetResource
 {
@@ -891,7 +934,12 @@ function Test-TargetResource
 
         [Parameter()]
         [System.Boolean]
-        $SuppressRestart
+        $SuppressRestart,
+
+        [Parameter()]
+        [ValidateSet('Mandatory', 'Optional', 'Strict')]
+        [System.String]
+        $Encrypt
     )
 
     $result = $true
@@ -900,6 +948,11 @@ function Test-TargetResource
         InstanceName         = $InstanceName
         DatabaseServerName   = $DatabaseServerName
         DatabaseInstanceName = $DatabaseInstanceName
+    }
+
+    if ($PSBoundParameters.ContainsKey('Encrypt'))
+    {
+        $getTargetResourceParameters.Encrypt = $Encrypt
     }
 
     $currentConfig = Get-TargetResource @getTargetResourceParameters